# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/waitlist": {
      "post": {
        "operationId": "joinWaitlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinWaitlistDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "waitlist"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumi API",
    "description": "Cuentas, sesión y lista de espera de Lumi.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "JoinWaitlistDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "app": {
            "type": "string",
            "description": "App que más le roba tiempo",
            "nullable": true
          },
          "form": {
            "type": "string",
            "nullable": true
          },
          "utm_source": {
            "type": "string",
            "nullable": true
          },
          "utm_medium": {
            "type": "string",
            "nullable": true
          },
          "utm_campaign": {
            "type": "string",
            "nullable": true
          },
          "utm_content": {
            "type": "string",
            "nullable": true
          },
          "ref": {
            "type": "string",
            "nullable": true
          },
          "referrer": {
            "type": "string",
            "nullable": true
          },
          "lang": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}